requirements.txt
matplotlib
neat-python
numpy
opencv-python
paho-mqtt
pygame
ultralytics